Update UI and Refactor
This cleans up a lot of the hacks that we did just to get things
going along with updating the styling to make the Messages page
at least look presentable
